+++ src/PhpPresentation/Writer/PowerPoint2007/Rels.php
@@ -53,6 +53,9 @@ class Rels extends AbstractPart
         $objWriter->startElement('Relationships');
         $objWriter->writeAttribute('xmlns', 'http://schemas.openxmlformats.org/package/2006/relationships');
 
+        // Relationship docProps/custom.xml
+        $this->writeRelationship($objWriter, 4, 'http://schemas.openxmlformats.org/officeDocument/2006/relationships/custom-properties', 'docProps/custom.xml');
+
         // Relationship docProps/app.xml
         $this->writeRelationship($objWriter, 3, 'http://schemas.openxmlformats.org/officeDocument/2006/relationships/extended-properties', 'docProps/app.xml');