+++ tests/PhpPresentation/Tests/Reader/PowerPoint2007Test.php
@@ -462,4 +462,20 @@ class PowerPoint2007Test extends \PHPUnit_Framework_TestCase
         $this->assertEquals('https://github.com/PHPOffice/PHPPresentation/', $oRichText->getHyperlink()->getUrl());
         $this->assertEquals('PHPPresentation', $oRichText->getHyperlink()->getTooltip());
     }
+
+    public function testMarkAsFinal()
+    {
+        $file = PHPPRESENTATION_TESTS_BASE_DIR . '/resources/files/Sample_12.pptx';
+        $object = new PowerPoint2007();
+        $oPhpPresentation = $object->load($file);
+        $this->assertInstanceOf('PhpOffice\\PhpPresentation\\PhpPresentation', $oPhpPresentation);
+        $this->assertFalse($oPhpPresentation->isMarkedAsFinal());
+
+
+        $file = PHPPRESENTATION_TESTS_BASE_DIR . '/resources/files/PPTX_MarkAsFinal.pptx';
+        $object = new PowerPoint2007();
+        $oPhpPresentation = $object->load($file);
+        $this->assertInstanceOf('PhpOffice\\PhpPresentation\\PhpPresentation', $oPhpPresentation);
+        $this->assertTrue($oPhpPresentation->isMarkedAsFinal());
+    }
 }